Removed relinkable handles from inflation-index interface.
Action at a distance (such as having the bootstrap process relink the handle
inside an index) can be a bit too surprising.  I uniformed the behavior with
that of the rest of the library---and in so doing, I broke a test that relied
on some hidden links.  The test (which exercises experimental stuff) has been
temporarily disabled, and the exercised classes have been marked as buggy.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk@16603 8618b1d8-e22c-0410-b026-96a5dba3e089