Fixed passing of arguments to FDMultiPeriodEngine (thanks to Selene Makarios.)
FDMultiPeriodEngine class was taking and passing the number of
time steps and grid points in the wrong order (the order of the
parameters in the signature is opposite to that in all other FD
classes.  For backward compatibility, this was not changed.)
A couple of dividend-option tests now require more time steps to pass.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk@17400 8618b1d8-e22c-0410-b026-96a5dba3e089