Renamed BaseIndexedCashFlow to less generic CPICashFlow.
The cash flow is currently pretty much tied to CPI instruments,
so I'd make this clear.  We'll need another use case to be able
to generalize.

Also, the name was somewhat confusing.  It was meant to convey
the idea of base date and base level, but instead it suggested
that it was a base class for indexed cash flows.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/trunk@18055 8618b1d8-e22c-0410-b026-96a5dba3e089